Add files via upload
The Fibonacci sequence is a sequence of numbers where each number is the sum of the two preceding numbers. In Python, this sequence can be generated using a loop or recursion. The Fibonacci sequence has many applications in mathematics and computer science, including in algorithms for searching and sorting. Also it is one of the most important concept in both programming and Mathematics. Mathematics is an important factor in Programming as well.